Record the requested bucket key for signups
You can't always get what you want, but if you try sometimes you get
into the anything bucket.  Or the waitlist.

This commit adds a column called `requested_bucket_key` to the
`signups` table, which records what bucket the user actually wanted
to sign up for.  This is actually what Intercode 1 records, and it
calculates the actual bucket the user ended up in at signup/drop
time (which isn't super hard to do, since the only possibilities are
Male, Female and Neutral).

In our more generalized algorithm, it may not always be possible in
the future to determine the actual bucket of a signup, and in any
case it will simplify a lot of display logic if we record it in the
database, so we'll record both the bucket the user wanted and the
bucket they actually got.