Bug fix: don't count temporary in-memory ticket for limit checks
While authorizing ProvideEventTicket, we temporarily create a Ticket
object in memory.  This didn't used to end up in the tickets
association cache, but at some point in Rails upgrades, it started doing
so.  This causes the ticket count check to be off by one when we
actually go to provide the ticket.

This works around the problem by reloading the event model (thereby
clearing its association caches) before trying to provide the ticket.