Disable invariant checks when rendering CMS pages
Our check for whether or not a template is invariant doesn't seem to be
fully working.  If a template is incorrectly marked as invariant, the
page it's on will effectively stop being updated as the underlying data
changes.  I'm not super convinced this optimization was doing a lot for
us anyway, so I'm simply going to disable it.