Fix "An object of type EventProposal" error
When using the Become User feature, admins may encounter the error
"An object of type EventProposal was hidden due to permissions".  This
is because only the user themselves is allowed to view draft event
proposals, even when using Become User.

This works around the issue by only returning event proposals that the
current session has permissions to view.